Fix loading of audio from glTF files
